re-worked cache to make it reusable; added cache for parsed queries, some javadoc, added shutdown hook to clean up cache
