+++ driver/src/test/java/com/strategicgains/docussandra/service/QueryServiceTest.java
@@ -94,91 +94,5 @@ public class QueryServiceTest
         assertEquals(expected, actual);
     }
 
-    /**
-     * Test of parseQuery method, of class QueryService.
-     */
-    @Test
-    public void testParseQueryBasic()
-    {
-        System.out.println("testParseQueryBasic");
-        String db = Fixtures.DB;
-        Query toParse = Fixtures.createTestQuery();
-        ParsedQuery expResult = new ParsedQuery(toParse, new WhereClause(toParse.getWhere()), "mydb_mytable_myindexwithonefield");
-        ParsedQuery result = instance.parseQuery(db, toParse);
-        assertEquals(expResult, result);
-    }
-
-    /**
-     * Test of parseQuery method, of class QueryService.
-     */
-    @Test
-    public void testParseQueryTwoFields()
-    {
-        System.out.println("testParseQueryTwoFields");
-        String db = Fixtures.DB;
-        Query toParse = Fixtures.createTestQuery2();
-        ParsedQuery expResult = new ParsedQuery(toParse, new WhereClause(toParse.getWhere()), "mydb_mytable_myindexwithtwofields");
-        ParsedQuery result = instance.parseQuery(db, toParse);
-        assertEquals(expResult, result);
-    }
-
-    /**
-     * Test of parseQuery method, of class QueryService.
-     */
-    @Test
-    public void testParseQueryTwoFieldsImperfectMatch()
-    {
-        System.out.println("testParseQueryTwoFieldsImperfectMatch");
-        String db = Fixtures.DB;
-        Query toParse = Fixtures.createTestQuery2();
-        toParse.setWhere("myindexedfield1 = 'thisismyfield'");
-        ParsedQuery expResult = new ParsedQuery(toParse, new WhereClause(toParse.getWhere()), "mydb_mytable_myindexwithtwofields");
-        ParsedQuery result = instance.parseQuery(db, toParse);
-        assertEquals(expResult, result);
-    }
-
-    /**
-     * Test of parseQuery method, of class QueryService.
-     */
-    @Test
-    public void testParseQueryException()
-    {
-        System.out.println("testParseQueryException");
-        String db = Fixtures.DB;
-        Query toParse = Fixtures.createTestQuery();
-        toParse.setWhere("nonIndexedField = 'boo'");
-        boolean expectedExceptionThrown = false;
-        try
-        {
-            ParsedQuery result = instance.parseQuery(db, toParse);
-        } catch (FieldNotIndexedException e)
-        {
-            expectedExceptionThrown = true;
-            assertTrue(e.getLocalizedMessage().contains("nonIndexedField"));
-        }
-        assertTrue("Expected exception not thrown", expectedExceptionThrown);
-    }
-
-    /**
-     * Test of equalLists method, of class QueryService.
-     */
-    @Test
-    public void testEqualLists()
-    {
-        System.out.println("equalLists");
-        List<String> one = new ArrayList<>();
-        List<String> two = new ArrayList<>();
-        boolean result = instance.equalLists(one, two);
-        assertEquals(true, result);
-        one.add("one");
-        one.add("two");
-        two.add("two");
-        two.add("one");
-        result = instance.equalLists(one, two);
-        assertEquals(true, result);
-        two.add("three");
-        result = instance.equalLists(one, two);
-        assertEquals(false, result);
-    }
-
+ 
 }
